SQL Server Leaderboard - revert to EFCore 1.1.0, implement default, top, aroundMe leaderboard

+++ src/Nether.Data/Leaderboard/GameScore.cs
@@ -9,7 +9,7 @@ namespace Nether.Data.Leaderboard
         public string Country { get; set; }
         public string CustomTag { get; set; }
         public int Score { get; set; }
-        public int Rank { get; set; }
+        public long Rank { get; set; }
     }
 }
